@@ -1,5 +1,5 @@
 <script>
-    export let data;
+    let { data } = $props();
 </script>
 
 <h1>buku.kresna.me</h1>
